refactor(storage): extract InteractionStoreMixin from ProfileMixin (SQLite + ABC)
Move the InteractionStore bucket (9 public methods + _insert_interaction)
verbatim out of the shrinking ProfileMixin into a new InteractionStoreMixin
under the profiles/ package, mirroring the Task-3 ProfileStore extraction.
Bodies are byte-identical; _insert_interaction keeps its commit-then-FTS/vec
ordering and the delete methods keep their inline interactions_fts/vec cleanup.
ProfileMixin now holds only interaction/profile search.
