+++ reflexio/server/services/storage/storage_base/profiles/__init__.py
@@ -1,5 +1,7 @@
+from ._interaction_store import InteractionStoreMixin
 from ._profile_store import ProfileStoreMixin
 
 __all__ = [
+    "InteractionStoreMixin",
     "ProfileStoreMixin",
 ]
